#4: Add S3 Integration and Swagger

+++ src/main/kotlin/ru/biluta/memes/service/persistence/model/MemesApp.kt
@@ -11,14 +11,16 @@ import java.time.LocalDateTime
 @Entity
 @Table(name="memes_app")
 class MemesApp (
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null,
     var chatId: Long,
-    var userId: Int,
+    var userId: Long,
     var author: String,
     var media: ByteArray,
     var chatPrefix: String,
     var mediaType: String,
     var createdAt: LocalDateTime = LocalDateTime.now()
+
 )
